requirements.txt
transformers==4.40.1
torch>=2.0.0
pytesseract==0.3.10
pdfplumber==0.10.2
python-docx==1.1.0
Pillow==10.3.0
customtkinter==5.2.1
tk
spacy==3.7.4
nltk==3.8.1
deep-translator==1.11.4
pyttsx3==2.90
fpdf==1.7.2
PyMuPDF==1.24.2
pywin32; platform_system=="Windows"
